fix: native support for retry template factory
